fix : 채팅 mysql redis 연동 및 기능 구현

+++ src/main/java/com/ssafy/homer/user/repository/UserRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 @Repository
 public interface UserRepository extends JpaRepository<User, Integer> {
 
-    Optional<User> findByUserId(Integer userId);
+    Optional<User> findByUserId(Long userId);
     
     Optional<User> findByEmail(String email);
 }
